Version 0.0.3
Dont know if im doing this right with Git but here are the Changes:
- CMSV2 has been Added and is now the standard Language, featuring:
-- C# Like Syntax
-- Native Fast C++ Conversion

- You can now share projects across machines, CTR will tell you that its corrupted, all you need to do is select the Directory of the Project and boom you can use it on another Device.

- Did a bit of Refining on the CTR Image Font Editor.
-- Whats the CTR Image Font Editor?
The CTR Image Font Editor is made to edit or make a CTR IF, a image font is just a series of images packed into a "Font" which will render from a string as seperate images per char.

CMSV2 Documentation and Syntax Coming soon.